port internal cursor stuff to gdouble, fix off-by-one in curve drawing,
2008-02-09  Michael Natterer  <mitch@gimp.org>

	* app/widgets/gimpcurveview.[ch]: port internal cursor stuff to
	gdouble, fix off-by-one in curve drawing, fix drawing artefact in
	handle drawing by starting drawing on the handle's outline and not
	its center.


svn path=/trunk/; revision=24839

+++ app/widgets/gimpcurveview.h
@@ -55,8 +55,8 @@ struct _GimpCurveView
   gdouble            xpos;
   PangoLayout       *xpos_layout;
 
-  gint               cursor_x;
-  gint               cursor_y;
+  gdouble            cursor_x;
+  gdouble            cursor_y;
   PangoLayout       *cursor_layout;
   PangoRectangle     cursor_rect;
 };
